gpu addition with new block and grid structure tested

+++ source/MatrixAlgorithms/array_algorithms.h
@@ -9,7 +9,7 @@
 int sum_of_array(int array[], int size_of_array);
 int *random_numbers(int number_count, int max_value);
 void print_numbers(int array[], int size_of_array);
-double mean(long array[], int size_of_array);
-double standard_deviation(long array[], int size_of_array, double mean);
+double mean(double array[], int size_of_array);
+double standard_deviation(double array[], int size_of_array, double mean);
 
 #endif
